+++ src/term.c
@@ -433,6 +433,7 @@ term_reset(bool full)
   term.suspend_update = 0;
   term.no_scroll = 0;
   term.scroll_mode = 0;
+  term.baud = cfg.baud;
 
   term_schedule_search_update();